refactor: remove slack org internal callback route
