+++ turbo/apps/api/src/lib/__tests__/internal-api-url.test.ts
@@ -13,10 +13,10 @@ describe("internalApiBaseUrl", () => {
     expect(internalApiBaseUrl()).toBe("https://api.vm0.ai");
     expect(
       new URL(
-        "/api/internal/callbacks/slack/org",
+        "/api/internal/callbacks/github/issues",
         internalApiBaseUrl(),
       ).toString(),
-    ).toBe("https://api.vm0.ai/api/internal/callbacks/slack/org");
+    ).toBe("https://api.vm0.ai/api/internal/callbacks/github/issues");
   });
 
   it("defaults to the API backend origin in production when VM0_API_BACKEND_URL is unset", () => {
